Move handler.go into gov.go
