+++ x/concentrated-liquidity/gov.go
@@ -1,7 +1,11 @@
 package concentrated_liquidity
 
 import (
+	"fmt"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
 	clmodel "github.com/osmosis-labs/osmosis/v16/x/concentrated-liquidity/model"
 	"github.com/osmosis-labs/osmosis/v16/x/concentrated-liquidity/types"
@@ -28,4 +32,22 @@ func (k Keeper) HandleCreateConcentratedLiquidityPoolsProposal(ctx sdk.Context, 
 // HandleTickSpacingDecreaseProposal handles a tick spacing decrease proposal to the corresponding keeper method.
 func (k Keeper) HandleTickSpacingDecreaseProposal(ctx sdk.Context, p *types.TickSpacingDecreaseProposal) error {
 	return k.DecreaseConcentratedPoolTickSpacing(ctx, p.PoolIdToTickSpacingRecords)
+}
+
+func NewConcentratedLiquidityProposalHandler(k Keeper) govtypes.Handler {
+	return func(ctx sdk.Context, content govtypes.Content) error {
+		switch c := content.(type) {
+		case *types.TickSpacingDecreaseProposal:
+			return handleTickSpacingDecreaseProposal(ctx, k, c)
+		case *types.CreateConcentratedLiquidityPoolsProposal:
+			return handleCreateConcentratedLiquidityPoolsProposal(ctx, k, c)
+
+		default:
+			return fmt.Errorf("unrecognized concentrated liquidity proposal content type: %T", c)
+		}
+	}
+}
+
+func handleTickSpacingDecreaseProposal(ctx sdk.Context, k Keeper, p *types.TickSpacingDecreaseProposal) error {
+	return k.HandleTickSpacingDecreaseProposal(ctx, p)
 }
