--- x/concentrated-liquidity/handler.go
@@ -1,28 +1,0 @@
-package concentrated_liquidity
-
-import (
-	"fmt"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-
-	"github.com/osmosis-labs/osmosis/v16/x/concentrated-liquidity/types"
-)
-
-func NewConcentratedLiquidityProposalHandler(k Keeper) govtypes.Handler {
-	return func(ctx sdk.Context, content govtypes.Content) error {
-		switch c := content.(type) {
-		case *types.TickSpacingDecreaseProposal:
-			return handleTickSpacingDecreaseProposal(ctx, k, c)
-		case *types.CreateConcentratedLiquidityPoolsProposal:
-			return handleCreateConcentratedLiquidityPoolsProposal(ctx, k, c)
-
-		default:
-			return fmt.Errorf("unrecognized concentrated liquidity proposal content type: %T", c)
-		}
-	}
-}
-
-func handleTickSpacingDecreaseProposal(ctx sdk.Context, k Keeper, p *types.TickSpacingDecreaseProposal) error {
-	return k.HandleTickSpacingDecreaseProposal(ctx, p)
-}
